+++ test/luciolucio/holi/add_test/days_test.clj
@@ -5,7 +5,7 @@
 
 (deftest should-calculate-correct-date-when-add-date-with-days
   (are [days expected]
-       (= expected (holi/add (t/date "2020-07-29") days :days))
+       (= expected (holi/add (t/date "2020-07-29") days :days) (holi/add (t/date "2020-07-29") days :day))
     0 (t/date "2020-07-29")
     1 (t/date "2020-07-30")
     11 (t/date "2020-08-09")
@@ -16,7 +16,7 @@
 
 (deftest should-calculate-correct-date-when-add-date-time-with-days
   (are [days expected]
-       (= expected (holi/add (t/date-time "2020-07-29T10:15:00") days :days))
+       (= expected (holi/add (t/date-time "2020-07-29T10:15:00") days :days) (holi/add (t/date-time "2020-07-29T10:15:00") days :day))
     0 (t/date-time "2020-07-29T10:15:00")
     1 (t/date-time "2020-07-30T10:15:00")
     11 (t/date-time "2020-08-09T10:15:00")