+++ test/luciolucio/holi/add_test/years_test.clj
@@ -5,7 +5,7 @@
 
 (deftest should-calculate-correct-date-when-add-date-with-years
   (are [days expected]
-       (= expected (holi/add (t/date "2020-02-29") days :years))
+       (= expected (holi/add (t/date "2020-02-29") days :years) (holi/add (t/date "2020-02-29") days :year))
     0 (t/date "2020-02-29")
     1 (t/date "2021-02-28")
     11 (t/date "2031-02-28")
@@ -16,7 +16,7 @@
 
 (deftest should-calculate-correct-date-when-add-date-time-with-years
   (are [days expected]
-       (= expected (holi/add (t/date-time "2020-02-29T10:15:00") days :years))
+       (= expected (holi/add (t/date-time "2020-02-29T10:15:00") days :years) (holi/add (t/date-time "2020-02-29T10:15:00") days :year))
     0 (t/date-time "2020-02-29T10:15:00")
     1 (t/date-time "2021-02-28T10:15:00")
     11 (t/date-time "2031-02-28T10:15:00")